ai-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.9.0
openai>=1.30.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
